Extract Fiber, FiberCache to new files
